i created the login and dashboard pages
