+++ src/app/features/admin/admin-routing.module.ts
@@ -2,15 +2,15 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { AdminComponent } from './admin.component';
-import { AuthComponent } from '../auth/auth.component';
-import { DashboardComponent } from '../dashboard/dashboard.component';
+import { AuthComponent } from '../auth/pages/page-auth/auth.component';
+import { DashboardComponent } from '../dashboard/pages/dashboard.component';
 
 const routes: Routes = [
   {
     path: '',
     component: AdminComponent,
     children: [
-      { path: '**', redirectTo: 'auth' },
+      { path: '*', redirectTo: '/admin/auth' },
       { path: 'auth', component: AuthComponent },
       { path: 'dashboard', component: DashboardComponent },
     ],
