--- src/app/features/auth/auth.component.html
@@ -1,3 +1,0 @@
-<p>
-  auth works!
-</p>
