+++ src/app/features/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AuthComponent } from './auth.component';
+import { AuthComponent } from './pages/page-auth/auth.component';
 
 @NgModule({
   imports: [
