+++ src/app/features/auth/pages/page-auth/auth.component.html
@@ -1,0 +1,25 @@
+<div class="text-center">
+<main class="form-signin">
+  <form>
+    <img class="mb-4" src="../assets/images/user.png" alt="" >
+    <h1 class="h3 mb-3 fw-normal">Ingreso al sistema</h1>
+
+    <div class="form-floating">
+      <input type="email" class="form-control" id="floatingInput" placeholder="name@example.com">
+      <label for="floatingInput">Email</label>
+    </div>
+    <div class="form-floating">
+      <input type="password" class="form-control" id="floatingPassword" placeholder="Password">
+      <label for="floatingPassword">Contraseña</label>
+    </div>
+
+    <div class="checkbox mb-3">
+      <label>
+        <input type="checkbox" value="remember-me"> Recordarme
+      </label>
+    </div>
+    <button (click)="login()" class="w-100 btn btn-lg btn-primary" type="button">Ingresar</button>
+    <p class="mt-5 mb-3 text-muted">&copy; 2017–2022</p>
+  </form>
+</main>
+</div>
