+++ src/app/features/auth/pages/page-auth/auth.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
+@Component({
+  selector: 'app-auth',
+  templateUrl: './auth.component.html',
+  styleUrls: ['./auth.component.css'],
+})
+export class AuthComponent implements OnInit {
+  constructor(private Router: Router) {}
+  login() {
+    this.Router.navigate(['admin/dashboard']);
+  }
+  ngOnInit() {}
+}
